pyproject.toml
[project]
name = "modbus-test"
version = "0.1.0"
description = "Modbus TCP performance testing tool"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pymodbus[tcp]>=3.0.0",
    "pyyaml>=6.0",
    "pandas>=1.5.0",
    "tabulate>=0.9.0"
]
